make base registerworkcommand
