+++ source/Octopus.Tentacle/Commands/RegisterWorkerCommand.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using Octopus.Client.Operations;
 using Octopus.Diagnostics;
 using Octopus.Tentacle.Commands.OptionSets;
@@ -11,40 +9,10 @@ using Octopus.Tentacle.Startup;
 
 namespace Octopus.Tentacle.Commands
 {
-    public class RegisterWorkerCommand : RegisterWorkerCommand<IRegisterWorkerOperation>
+    public class RegisterWorkerCommand : RegisterWorkerCommandBase<IRegisterWorkerOperation>
     {
         public RegisterWorkerCommand(Lazy<IRegisterWorkerOperation> lazyRegisterMachineOperation, Lazy<IWritableTentacleConfiguration> configuration, ISystemLog log, IApplicationInstanceSelector selector, Lazy<IOctopusServerChecker> octopusServerChecker, IProxyConfigParser proxyConfig, IOctopusClientInitializer octopusClientInitializer, ISpaceRepositoryFactory spaceRepositoryFactory, ILogFileOnlyLogger logFileOnlyLogger) : base(lazyRegisterMachineOperation, configuration, log, selector, octopusServerChecker, proxyConfig, octopusClientInitializer, spaceRepositoryFactory, logFileOnlyLogger)
         {
-        }
-    }
-    
-    public class RegisterWorkerCommand<TRegisterWorkerOperation> : RegisterMachineCommandBase<TRegisterWorkerOperation> where TRegisterWorkerOperation : IRegisterWorkerOperation
-    {
-        readonly List<string> workerpools = new List<string>();
-
-        public RegisterWorkerCommand(Lazy<TRegisterWorkerOperation> lazyRegisterMachineOperation,
-            Lazy<IWritableTentacleConfiguration> configuration,
-            ISystemLog log,
-            IApplicationInstanceSelector selector,
-            Lazy<IOctopusServerChecker> octopusServerChecker,
-            IProxyConfigParser proxyConfig,
-            IOctopusClientInitializer octopusClientInitializer,
-            ISpaceRepositoryFactory spaceRepositoryFactory,
-            ILogFileOnlyLogger logFileOnlyLogger)
-            : base(lazyRegisterMachineOperation, configuration, log, selector, octopusServerChecker, proxyConfig, octopusClientInitializer, spaceRepositoryFactory, logFileOnlyLogger)
-        {
-            Options.Add("workerpool=", "The worker pool name, slug or Id to add the machine to - e.g., 'Windows Pool'; specify this argument multiple times to add to multiple pools", s => workerpools.Add(s));
-        }
-
-        protected override void CheckArgs()
-        {
-            if (workerpools.Count == 0 || string.IsNullOrWhiteSpace(workerpools.First()))
-                throw new ControlledFailureException("Please specify a worker pool name, slug or Id, e.g., --workerpool=Default");
-        }
-
-        protected override void EnhanceOperation(TRegisterWorkerOperation registerOperation)
-        {
-            registerOperation.WorkerPools = workerpools.ToArray();
         }
     }
 }
